Commit the changes of CourseController, CourseService and CourseServiceImpl to update Delete function.

+++ src/main/java/com/cydeo/controller/CourseController.java
@@ -8,6 +8,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
 
 @Controller
 @RequestMapping("/course")
@@ -31,7 +33,7 @@ public class CourseController {
 
     @PostMapping("/create")
     public String insertCourse(@Valid @ModelAttribute("course") Course course, BindingResult bindingResult, Model model) {
-        if(bindingResult.hasErrors()){
+        if (bindingResult.hasErrors()) {
             model.addAttribute("managers", userService.findManagers());
             model.addAttribute("courses", courseService.findAll());
             return "/course/course-create";
@@ -44,16 +46,37 @@ public class CourseController {
 
 
     //update
-    @PostMapping("/update")
-    public String updateCourse(@ModelAttribute("course") Course course){
+
+    @GetMapping("/update/{id}")
+    public String editCourse(@PathVariable("id") Long id, Model model) {
+        model.addAttribute("course", courseService.findById(id));
+        model.addAttribute("managers", userService.findManagers());
+        model.addAttribute("courses", courseService.findAll());
+        return "/course/course-update";
+    }
+
+    @PostMapping("/update/{id}")
+    public String updateCourse(@Valid @ModelAttribute("course") Course course, BindingResult bindingResult, Model model) {
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("managers", userService.findManagers());
+            model.addAttribute("courses", courseService.findAll());
+            return "/course/course-update";
+        }
         courseService.update(course);
         return "redirect:/course/create";
     }
 
     //delete
+
     @GetMapping("/delete/{id}")
-    public String deleteCourse(@PathVariable("id") Long id) {
-        courseService.deleteById(id);
+    public String deleteCourse(@PathVariable("id") Long id, RedirectAttributes redirectAttributes) {
+
+        if (courseService.isNotEligibleToDelete(id)) {
+            redirectAttributes.addFlashAttribute("error", "This Course has either one or more than one lessons. Not allowed to delete");
+        } else {
+            redirectAttributes.addFlashAttribute("success", "This Course is successfully deleted");
+            courseService.deleteById(id);
+        }
         return "redirect:/course/create";
     }
